refactor(classwork): separate modal presentation and lifecycles
