+++ src/components/surveys/TeacherSurveyWorkspace.tsx
@@ -14,7 +14,7 @@ import { Code, ExternalLink, Eye, Plus, RotateCcw, Trash2 } from 'lucide-react'
 import { addDaysToDateString } from '@/lib/date-string'
 import { Button, Card, ConfirmDialog, FormField, Input, Select } from '@/ui'
 import { AssessmentSetupCheckbox } from '@/components/assessment/AssessmentSetupForm'
-import { DateTimeFields } from '@/components/classwork/ClassworkContentModal'
+import { DateTimeFields } from '@/components/classwork/ClassworkDueFields'
 import { EditableAssessmentTitle } from '@/components/assessment/EditableAssessmentTitle'
 import { QuestionMarkdown } from '@/components/QuestionMarkdown'
 import { Spinner } from '@/components/Spinner'
