+++ tests/api/teacher/surveys-route.test.ts
@@ -833,29 +833,16 @@ describe('POST /api/teacher/surveys', () => {
     })
   })
 
-  it('retries survey creation without due fields when the due columns are not migrated yet', async () => {
-    const warnSpy = vi.spyOn(console, 'warn').mockImplementation(() => {})
+  it('requires migration instead of dropping due fields during survey creation', async () => {
     const insert = vi.fn((payload: Record<string, unknown>) => ({
       select: vi.fn(() => ({
-        single: vi.fn().mockResolvedValueOnce(
-          insert.mock.calls.length === 1
-            ? {
-                data: null,
-                error: {
-                  code: 'PGRST204',
-                  message: "Could not find the 'due_at' column of 'surveys' in the schema cache",
-                },
-              }
-            : {
-                data: {
-                  id: 'survey-1',
-                  status: 'draft',
-                  opens_at: null,
-                  ...payload,
-                },
-                error: null,
-              },
-        ),
+        single: vi.fn().mockResolvedValue({
+          data: null,
+          error: {
+            code: 'PGRST204',
+            message: "Could not find the 'due_at' column of 'surveys' in the schema cache",
+          },
+        }),
       })),
     }))
     ;(mockSupabaseClient.from as any) = vi.fn((table: string) => {
@@ -915,20 +902,16 @@ describe('POST /api/teacher/surveys', () => {
     )
     const data = await response.json()
 
-    expect(response.status).toBe(201)
-    expect(insert).toHaveBeenCalledTimes(2)
+    expect(response.status).toBe(503)
+    expect(insert).toHaveBeenCalledTimes(1)
     expect(insert.mock.calls[0][0]).toEqual(expect.objectContaining({
       due_at: '2026-01-02T20:30:00.000Z',
       due_policy: 'soft',
     }))
-    expect(insert.mock.calls[1][0]).not.toHaveProperty('due_at')
-    expect(insert.mock.calls[1][0]).not.toHaveProperty('due_policy')
-    expect(data.survey).toEqual(expect.objectContaining({
-      id: 'survey-1',
-      title: 'Weekly check-in',
-      due_at: null,
-      due_policy: 'soft',
-    }))
-    warnSpy.mockRestore()
+    expect(data).toEqual({
+      error: 'Survey due dates are unavailable until migration 080 is applied.',
+      code: 'SURVEY_DUE_MIGRATION_REQUIRED',
+      migration_required: true,
+    })
   })
 })
