+++ tests/components/TeacherClassroomView.test.tsx
@@ -1535,6 +1535,8 @@ describe('TeacherClassroomView', () => {
     await within(materialDialog).findByRole('button', { name: 'Post Material' })
 
     expect(within(materialDialog).getByText('Saved')).toBeInTheDocument()
+    expect(within(materialDialog).getByPlaceholderText('Reading, link, handout...')).toBeInTheDocument()
+    expect(within(materialDialog).getByRole('button', { name: 'Schedule...' })).toBeInTheDocument()
     expect(within(materialDialog).queryByRole('button', { name: 'Delete' })).not.toBeInTheDocument()
     expect(within(materialDialog).queryByRole('button', { name: 'Save Draft' })).not.toBeInTheDocument()
     expect(global.fetch).toHaveBeenCalledWith(
